// src/api/types.gen.ts
export type ContactUsStatus = 'PENDING' | 'READ' | 'RESPONDED' | 'ARCHIVED';

export type ContactUs = {
    id: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    name: string;
    email: string;
    subject: string;
    message: string;
    status: ContactUsStatus;
};